Refactor backend for SOLID compliance
Centralized token logic with ITokenService, removing duplicate methods from multiple controllers. Extracted session activation logic into ISessionActivationService, replacing redundant loops in AdminController and SessionsController. Removed MockStripeService in favor of MockPaymentService.

Moved 7 inline request models from AdminController to individual files under Models/Requests. Updated controllers to use the new services and models. Registered ITokenService and ISessionActivationService in Program.cs.

Added backend-refactor.json to document the phased refactor plan, including SOLID violations, target structure, and testing checkpoints. Verified all tests pass and functionality remains intact.
