+++ HockeyRinkAPI/Models/Requests/PublishDraftRequest.cs
@@ -1,0 +1,6 @@
+namespace HockeyRinkAPI.Models.Requests;
+
+public class PublishDraftModel
+{
+    public bool Published { get; set; }
+}
