+++ HockeyRinkAPI/Services/ISessionActivationService.cs
@@ -1,0 +1,8 @@
+using HockeyRinkAPI.Models;
+
+namespace HockeyRinkAPI.Services;
+
+public interface ISessionActivationService
+{
+    Task<bool> ApplyActivationRulesAsync(IEnumerable<Session> sessions);
+}
