+++ HockeyRinkAPI/Services/ITokenService.cs
@@ -1,0 +1,10 @@
+using HockeyRinkAPI.Models;
+
+namespace HockeyRinkAPI.Services;
+
+public interface ITokenService
+{
+    string GenerateToken(ApplicationUser user);
+    Task<bool> ValidateTokenAsync(string token);
+    Task<string?> GetUserIdFromTokenAsync(string token);
+}
